requirements.txt
asttokens==2.4.1
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
executing==2.0.1
fabric==3.2.2
Fabric3==1.14.post1
Flask==3.0.3
Flask-API==3.1
greenlet==3.0.3
idna==3.7
invoke==2.2.0
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mysqlclient==2.2.4
nest-asyncio==1.6.0
packaging==24.1
paramiko==2.12.0
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.22
Pygments==2.18.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
pyzmq==26.0.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.30
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
wrapt==1.16.0
